コンパイル時に起きていたエラーになりうるものを少し消した

+++ Engine/Math/Vector3.cpp
@@ -45,9 +45,9 @@ Vector3 Vector3::cross(const Vector3& v) const
 Vector3 Vector3::distanceFrom(const Vector3& v, const Vector3& v2) const
 {
 	return  Vector3(
-		std::sqrt(std::pow(v2.x - v.x, 2)),
-		std::sqrt(std::pow(v2.y - v.y, 2)),
-		std::sqrt(std::pow(v2.z - v.z, 2))
+		(float)std::sqrt(std::pow(v2.x - v.x, 2)),
+		(float)std::sqrt(std::pow(v2.y - v.y, 2)),
+		(float)std::sqrt(std::pow(v2.z - v.z, 2))
 	);
 }
